Move testTimeZoneId to ItSqlApiBaseTest
